+++ wallettemplate/src/main/java/wallettemplate/Main.java
@@ -30,21 +30,17 @@ import org.bitcoinj.wallet.DeterministicSeed;
 import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.layout.Pane;
-import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
-import wallettemplate.controls.NotificationBarPane;
 import org.bitcoinj.walletfx.utils.GuiUtils;
-import org.bitcoinj.walletfx.utils.TextFieldValidator;
 
 import javax.annotation.Nullable;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 
-import static org.bitcoinj.walletfx.utils.GuiUtils.*;
+import static org.bitcoinj.walletfx.utils.GuiUtils.informationalAlert;
 
 public class Main extends Application {
     public static NetworkParameters params = TestNet3Params.get();
@@ -56,11 +52,7 @@ public class Main extends Application {
     public static WalletAppKit bitcoin;
     public static Main instance;
 
-    private StackPane uiStack;
-    private Pane mainUI;
     public MainController controller;
-    public NotificationBarPane notificationBar;
-    public Stage mainWindow;
 
     @Override
     public void start(Stage mainWindow) throws Exception {
@@ -73,7 +65,6 @@ public class Main extends Application {
     }
 
     private void realStart(Stage mainWindow) throws IOException {
-        this.mainWindow = mainWindow;
         instance = this;
         // Show the crash dialog for any exceptions that we don't handle and that hit the main loop.
         GuiUtils.handleCrashesOnThisThread();
@@ -87,19 +78,10 @@ public class Main extends Application {
         // Load the GUI. The MainController class will be automagically created and wired up.
         URL location = getClass().getResource("main.fxml");
         FXMLLoader loader = new FXMLLoader(location);
-        mainUI = loader.load();
+        Pane mainUI = loader.load();
         controller = loader.getController();
-        // Configure the window with a StackPane so we can overlay things on top of the main UI, and a
-        // NotificationBarPane so we can slide messages and progress bars in from the bottom. Note that
-        // ordering of the construction and connection matters here, otherwise we get (harmless) CSS error
-        // spew to the logs.
-        notificationBar = new NotificationBarPane(mainUI);
-        mainWindow.setTitle(APP_NAME);
-        uiStack = new StackPane();
-        Scene scene = new Scene(uiStack);
-        TextFieldValidator.configureScene(scene);   // Add CSS that we need.
-        scene.getStylesheets().add(getClass().getResource("wallet.css").toString());
-        uiStack.getChildren().add(notificationBar);
+
+        Scene scene = controller.controllerStart(mainUI, "wallet.css");
         mainWindow.setScene(scene);
 
         // Make log output concise.
@@ -154,94 +136,6 @@ public class Main extends Application {
             bitcoin.restoreWalletFromSeed(seed);
     }
 
-    private Node stopClickPane = new Pane();
-
-    public class OverlayUI<T> {
-        public Node ui;
-        public T controller;
-
-        public OverlayUI(Node ui, T controller) {
-            this.ui = ui;
-            this.controller = controller;
-        }
-
-        public void show() {
-            checkGuiThread();
-            if (currentOverlay == null) {
-                uiStack.getChildren().add(stopClickPane);
-                uiStack.getChildren().add(ui);
-                blurOut(mainUI);
-                //darken(mainUI);
-                fadeIn(ui);
-                zoomIn(ui);
-            } else {
-                // Do a quick transition between the current overlay and the next.
-                // Bug here: we don't pay attention to changes in outsideClickDismisses.
-                explodeOut(currentOverlay.ui);
-                fadeOutAndRemove(uiStack, currentOverlay.ui);
-                uiStack.getChildren().add(ui);
-                ui.setOpacity(0.0);
-                fadeIn(ui, 100);
-                zoomIn(ui, 100);
-            }
-            currentOverlay = this;
-        }
-
-        public void outsideClickDismisses() {
-            stopClickPane.setOnMouseClicked((ev) -> done());
-        }
-
-        public void done() {
-            checkGuiThread();
-            if (ui == null) return;  // In the middle of being dismissed and got an extra click.
-            explodeOut(ui);
-            fadeOutAndRemove(uiStack, ui, stopClickPane);
-            blurIn(mainUI);
-            //undark(mainUI);
-            this.ui = null;
-            this.controller = null;
-            currentOverlay = null;
-        }
-    }
-
-    @Nullable
-    private OverlayUI currentOverlay;
-
-    public <T> OverlayUI<T> overlayUI(Node node, T controller) {
-        checkGuiThread();
-        OverlayUI<T> pair = new OverlayUI<>(node, controller);
-        // Auto-magically set the overlayUI member, if it's there.
-        try {
-            controller.getClass().getField("overlayUI").set(controller, pair);
-        } catch (IllegalAccessException | NoSuchFieldException ignored) {
-        }
-        pair.show();
-        return pair;
-    }
-
-    /** Loads the FXML file with the given name, blurs out the main UI and puts this one on top. */
-    public <T> OverlayUI<T> overlayUI(String name) {
-        try {
-            checkGuiThread();
-            // Load the UI from disk.
-            URL location = GuiUtils.getResource(name);
-            FXMLLoader loader = new FXMLLoader(location);
-            Pane ui = loader.load();
-            T controller = loader.getController();
-            OverlayUI<T> pair = new OverlayUI<>(ui, controller);
-            // Auto-magically set the overlayUI member, if it's there.
-            try {
-                if (controller != null)
-                    controller.getClass().getField("overlayUI").set(controller, pair);
-            } catch (IllegalAccessException | NoSuchFieldException ignored) {
-                ignored.printStackTrace();
-            }
-            pair.show();
-            return pair;
-        } catch (IOException e) {
-            throw new RuntimeException(e);  // Can't happen.
-        }
-    }
 
     @Override
     public void stop() throws Exception {
