+++ wallettemplate/src/main/java/wallettemplate/MainController.java
@@ -19,6 +19,12 @@ package wallettemplate;
 import javafx.beans.binding.Binding;
 import javafx.beans.binding.Bindings;
 import javafx.beans.value.ObservableValue;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
+import javafx.scene.Scene;
+import javafx.scene.input.KeyCombination;
+import javafx.scene.layout.Pane;
+import javafx.scene.layout.StackPane;
 import org.bitcoinj.core.listeners.DownloadProgressTracker;
 import org.bitcoinj.core.Coin;
 import org.bitcoinj.utils.MonetaryFormat;
@@ -30,12 +36,19 @@ import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.util.Duration;
+import org.bitcoinj.walletfx.utils.GuiUtils;
+import org.bitcoinj.walletfx.utils.TextFieldValidator;
 import wallettemplate.controls.ClickableBitcoinAddress;
 import wallettemplate.controls.NotificationBarPane;
 import org.bitcoinj.walletfx.utils.BitcoinUIModel;
 import org.bitcoinj.walletfx.utils.easing.EasingMode;
 import org.bitcoinj.walletfx.utils.easing.ElasticInterpolator;
 
+import javax.annotation.Nullable;
+import java.io.IOException;
+import java.net.URL;
+
+import static org.bitcoinj.walletfx.utils.GuiUtils.*;
 import static wallettemplate.Main.bitcoin;
 
 /**
@@ -43,18 +56,42 @@ import static wallettemplate.Main.bitcoin;
  * after. This class handles all the updates and event handling for the main UI.
  */
 public class MainController {
+    public static MainController instance;
     public HBox controlsBox;
     public Label balance;
     public Button sendMoneyOutBtn;
     public ClickableBitcoinAddress addressControl;
 
-    private BitcoinUIModel model = new BitcoinUIModel();
+    private final BitcoinUIModel model = new BitcoinUIModel();
     private NotificationBarPane.Item syncItem;
     private static final MonetaryFormat MONETARY_FORMAT = MonetaryFormat.BTC.noCode();
 
+    private Pane mainUI;
+    private StackPane uiStack;
+    private NotificationBarPane notificationBar;
+    private final Node stopClickPane = new Pane();
+
     // Called by FXMLLoader.
     public void initialize() {
+        instance = this;
         addressControl.setOpacity(0.0);
+    }
+
+    Scene controllerStart(Pane mainUI, String cssResourceName) {
+        this.mainUI = mainUI;
+        // Configure the window with a StackPane so we can overlay things on top of the main UI, and a
+        // NotificationBarPane so we can slide messages and progress bars in from the bottom. Note that
+        // ordering of the construction and connection matters here, otherwise we get (harmless) CSS error
+        // spew to the logs.
+        notificationBar = new NotificationBarPane(mainUI);
+        uiStack = new StackPane();
+        Scene scene = new Scene(uiStack);
+        TextFieldValidator.configureScene(scene);
+        // Add CSS that we need. cssResourceName will be loaded from the same package as this class.
+        scene.getStylesheets().add(getClass().getResource(cssResourceName).toString());
+        uiStack.getChildren().add(notificationBar);
+        scene.getAccelerators().put(KeyCombination.valueOf("Shortcut+F"), () -> bitcoin.peerGroup().getDownloadPeer().close());
+        return scene;
     }
 
     public void onBitcoinSetup() {
@@ -87,16 +124,16 @@ public class MainController {
     }
 
     private void showBitcoinSyncMessage() {
-        syncItem = Main.instance.notificationBar.pushItem("Synchronising with the Bitcoin network", model.syncProgressProperty());
+        syncItem = notificationBar.pushItem("Synchronising with the Bitcoin network", model.syncProgressProperty());
     }
 
     public void sendMoneyOut(ActionEvent event) {
         // Hide this UI and show the send money UI. This UI won't be clickable until the user dismisses send_money.
-        Main.instance.overlayUI("send_money.fxml");
+        overlayUI("send_money.fxml");
     }
 
     public void settingsClicked(ActionEvent event) {
-        Main.OverlayUI<WalletSettingsController> screen = Main.instance.overlayUI("wallet_settings.fxml");
+        OverlayUI<WalletSettingsController> screen = overlayUI("wallet_settings.fxml");
         screen.controller.initialize(null);
     }
 
@@ -122,5 +159,92 @@ public class MainController {
 
     public DownloadProgressTracker progressBarUpdater() {
         return model.getDownloadProgressTracker();
+    }
+
+    public class OverlayUI<T> {
+        public Node ui;
+        public T controller;
+
+        public OverlayUI(Node ui, T controller) {
+            this.ui = ui;
+            this.controller = controller;
+        }
+
+        public void show() {
+            checkGuiThread();
+            if (currentOverlay == null) {
+                uiStack.getChildren().add(stopClickPane);
+                uiStack.getChildren().add(ui);
+                blurOut(mainUI);
+                //darken(mainUI);
+                fadeIn(ui);
+                zoomIn(ui);
+            } else {
+                // Do a quick transition between the current overlay and the next.
+                // Bug here: we don't pay attention to changes in outsideClickDismisses.
+                explodeOut(currentOverlay.ui);
+                fadeOutAndRemove(uiStack, currentOverlay.ui);
+                uiStack.getChildren().add(ui);
+                ui.setOpacity(0.0);
+                fadeIn(ui, 100);
+                zoomIn(ui, 100);
+            }
+            currentOverlay = this;
+        }
+
+        public void outsideClickDismisses() {
+            stopClickPane.setOnMouseClicked((ev) -> done());
+        }
+
+        public void done() {
+            checkGuiThread();
+            if (ui == null) return;  // In the middle of being dismissed and got an extra click.
+            explodeOut(ui);
+            fadeOutAndRemove(uiStack, ui, stopClickPane);
+            blurIn(mainUI);
+            //undark(mainUI);
+            this.ui = null;
+            this.controller = null;
+            currentOverlay = null;
+        }
+    }
+
+    @Nullable
+    private OverlayUI currentOverlay;
+
+    public <T> OverlayUI<T> overlayUI(Node node, T controller) {
+        checkGuiThread();
+        OverlayUI<T> pair = new OverlayUI<>(node, controller);
+        // Auto-magically set the overlayUI member, if it's there.
+        try {
+            controller.getClass().getField("overlayUI").set(controller, pair);
+        } catch (IllegalAccessException | NoSuchFieldException ignored) {
+        }
+        pair.show();
+        return pair;
+    }
+
+    /** Loads the FXML file with the given name, blurs out the main UI and puts this one on top. */
+    public <T> OverlayUI<T> overlayUI(String name) {
+        try {
+            checkGuiThread();
+            // Load the UI from disk.
+            URL location = GuiUtils.getResource(name);
+            FXMLLoader loader = new FXMLLoader(location);
+            Pane ui = loader.load();
+            T controller = loader.getController();
+            OverlayUI<T> pair = new OverlayUI<>(ui, controller);
+            // Auto-magically set the overlayUI member, if it's there.
+            try {
+                if (controller != null)
+                    controller.getClass().getField("overlayUI").set(controller, pair);
+            } catch (IllegalAccessException | NoSuchFieldException ignored) {
+                ignored.printStackTrace();
+            }
+            pair.show();
+            return pair;
+        } catch (IOException e) {
+            throw new RuntimeException(e);  // Can't happen.
+        }
     }
 }
